initialize transposition table

+++ src/evaluate.cpp
@@ -3,9 +3,9 @@
 
 using namespace std;
 
-double evaluate(thc::ChessRules &board)
+float evaluate(thc::ChessRules &board)
 {
-    double evaluation = 0;
+    float evaluation = 0;
 
     // Queens alive?
 
@@ -24,15 +24,15 @@ double evaluate(thc::ChessRules &board)
 
                 if (rank < 5 && (file == 4 || file == 3))
                 {
-                    evaluation += 0.4;
+                    evaluation += 0.4f;
                 }
                 if (rank < 6 && (board.squares[(rank + 1) * 8 + file + 1] == 'P' || board.squares[(rank + 1 * 8) + file - 1] == 'P'))
                 {
-                    evaluation += 0.3;
+                    evaluation += 0.3f;
                 }
                 if (board.squares[(rank + 1) * 8 + file] == 'P')
                 {
-                    evaluation -= 0.19;
+                    evaluation -= 0.19f;
                 }
 
                 break;
@@ -43,17 +43,17 @@ double evaluate(thc::ChessRules &board)
 
                 if (rank > 2 && (file == 4 || file == 3))
                 {
-                    evaluation -= 0.4;
+                    evaluation -= 0.4f;
                 }
                 //likes pawns being protected by other pawns maybe?
                 if (rank > 1 && (board.squares[(rank - 1) * 8 + file + 1] == 'p' || board.squares[(rank - 1) * 8 + file - 1] == 'p'))
                 {
-                    evaluation -= 0.3;
+                    evaluation -= 0.3f;
                 }
                 //doesnt like doubled pawns
                 if (board.squares[(rank - 1) * 8 + file] == 'p')
                 {
-                    evaluation += 0.19;
+                    evaluation += 0.19f;
                 }
 
                 break;
@@ -64,11 +64,11 @@ double evaluate(thc::ChessRules &board)
 
                 if (rank == 7)
                 {
-                    evaluation -= 0.33;
+                    evaluation -= 0.33f;
                 }
                 if (file == 0 || file == 7)
                 {
-                    evaluation -= 0.5;
+                    evaluation -= 0.5f;
                 }
 
                 break;
@@ -79,41 +79,41 @@ double evaluate(thc::ChessRules &board)
 
                 if (rank == 0)
                 {
-                    evaluation += 0.33;
+                    evaluation += 0.33f;
                 }
                 if (file == 0 || file == 7)
                 {
-                    evaluation += 0.5;
+                    evaluation += 0.5f;
                 }
 
                 break;
             }
             case 'B':
             {
-                evaluation += 3.2;
+                evaluation += 3.2f;
 
                 if (rank == 7)
                 {
-                    evaluation -= 0.2;
+                    evaluation -= 0.2f;
                 }
                 if (rank == 6 && (file == 1 || file == 6))
                 {
-                    evaluation += 0.35;
+                    evaluation += 0.35f;
                 }
 
                 break;
             }
             case 'b':
             {
-                evaluation -= 3.2;
+                evaluation -= 3.2f;
 
                 if (rank == 0)
                 {
-                    evaluation += 0.2;
+                    evaluation += 0.2f;
                 }
                 if (rank == 1 && (file == 1 || file == 6))
                 {
-                    evaluation -= 0.35;
+                    evaluation -= 0.35f;
                 }
 
                 break;
@@ -124,16 +124,16 @@ double evaluate(thc::ChessRules &board)
 
                 if (file == 0 || file == 1 || file == 6 || file == 7)
                 {
-                    evaluation -= 0.8;
+                    evaluation -= 0.8f;
                 }
                 if (!(rank == 0 || rank == 1 || rank == 6 || rank == 7))
                 {
-                    evaluation -= 0.22;
+                    evaluation -= 0.22f;
                 }
                 // likes rooks next to each other
                 if (!((rank * 8 + file + 1) == 'R' || (rank * 8 + file + -1) == 'R'))
                 {
-                    evaluation -= 0.3;
+                    evaluation -= 0.3f;
                 }
 
                 break;
@@ -144,16 +144,16 @@ double evaluate(thc::ChessRules &board)
 
                 if (file == 0 || file == 1 || file == 6 || file == 7)
                 {
-                    evaluation += 0.8;
+                    evaluation += 0.8f;
                 }
                 if (!(rank == 0 || rank == 1 || rank == 6 || rank == 7))
                 {
-                    evaluation += 0.22;
+                    evaluation += 0.22f;
                 }
                 // likes rooks next to each other
                 if (!((rank * 8 + file + 1) == 'r' || (rank * 8 + file + -1) == 'r'))
                 {
-                    evaluation += 0.3;
+                    evaluation += 0.3f;
                 }
 
                 break;
@@ -175,7 +175,7 @@ double evaluate(thc::ChessRules &board)
                 //check if the king is safe when castled
                 if (((rank - 1) * 8 + file == 'P') && ((rank - 1) * 8 + file + 1 == 'P'))
                 {
-                    evaluation += 0.3;
+                    evaluation += 0.3f;
                 }
                 break;
             }
@@ -183,7 +183,7 @@ double evaluate(thc::ChessRules &board)
             {
                 if (((rank + 1) * 8 + file == 'p') && ((rank + 1) * 8 + file + 1 == 'p'))
                 {
-                    evaluation -= 0.3;
+                    evaluation -= 0.3f;
                 }
                 break;
             }
@@ -193,12 +193,12 @@ double evaluate(thc::ChessRules &board)
 
     if ((board.squares[58] == 'K' || board.squares[62] == 'K') && wkingSafety == false)
     {
-        evaluation += 0.5;
+        evaluation += 0.5f;
     }
 
     if ((board.squares[2] == 'k' || board.squares[6] == 'k') && bkingSafety == false)
     {
-        evaluation -= 0.5;
+        evaluation -= 0.5f;
     }
 
     evaluation += (rand() % 2 - 1) / 100;
