+++ src/search.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "../lib/thc/thc.h"
 
-void search(thc::ChessRules board, int maxDepth);
+void search(thc::ChessRules board, short int maxDepth);
